+++ assets/project-overview.js
@@ -261,11 +261,15 @@ function changeMilestone(event, ticketId, newMilestoneId) {
         const newMilestoneColor = jQuery(
           `#milestone-option-${newMilestoneId}`
         ).attr('data-color');
-        const isItAHexColor = hexColorRegExp.exec(newMilestoneColor);
 
+        const isItAHexColor = hexColorRegExp.exec(newMilestoneColor);
         if (isItAHexColor) {
           jQuery(parentElement).css('background', newMilestoneColor);
+          jQuery('#milestone-select').addClass('milestone-select-white-text');
         } else {
+          jQuery('#milestone-select').removeClass(
+            'milestone-select-white-text'
+          );
           jQuery(parentElement).css('background', 'transparent');
         }
       });